CCM-8993: rename to preview digital
