+++ frontend/src/components/organisms/PreviewDigitalTemplate/PreviewDigitalTemplate.tsx
@@ -2,7 +2,7 @@
 
 import { ZodErrorSummary } from '@molecules/ZodErrorSummary/ZodErrorSummary';
 import { NHSNotifyRadioButtonForm } from '@molecules/NHSNotifyRadioButtonForm/NHSNotifyRadioButtonForm';
-import { PreviewTemplateProps } from './preview-template.types';
+import { PreviewTemplateProps } from './preview-digitial-template.types';
 
 export function PreviewTemplate({
   form,
