+++ frontend/src/components/organisms/PreviewDigitalTemplate/index.ts
@@ -1,0 +1,2 @@
+export * from './PreviewDigitalTemplate';
+export * from './preview-digitial-template.types';
