+++ frontend/src/components/organisms/PreviewDigitalTemplate/preview-digitial-template.types.ts
@@ -5,6 +5,7 @@ import { TemplateDto } from 'nhs-notify-backend-client';
 export type PreviewTemplateProps = {
   sectionHeading: string | undefined;
   template: TemplateDto;
+  
   form: {
     errorHeading: string;
   } & NHSNotifyRadioButtonFormProps;
