--- frontend/src/components/organisms/PreviewTemplate/index.ts
@@ -1,2 +1,0 @@
-export * from './PreviewTemplate';
-export * from './preview-template.types';
